requirement.txt
streamlit
openai
sentence-transformers
nltk
numpy
PyPDF2
python-docx